Adding a `text_to_instance` method on DatasetReader (#256)
* Adding a `text_to_instance` method on DatasetReader

This is so that the `Predictors`, or anything else that wants to make
predictions given a trained model, has an easier time of processing the
data in the same way the trained model did - all you need is the
instantiated `DatasetReader`, and not any private members pulled out
from that class.

Changing this on the `DatasetReaders`, though, meant changing the
`Predictors` themselves.  And that meant unifying the API around making
model predictions from single instance inputs.  I was able to remove
all of the `tag` and `predict_` methods on individual models, replacing
them with a single `Model.forward_on_instance` method.  I _think_ the
result is a lot cleaner, but I could have missed something important
when making this change - it was a whole bunch of pulling things apart
and putting them back together.

* Use text_to_instance in _process_sentence

* Moved decoding logic back into models

* Fix mypy and pylint