Rename EncoderDecoder to SimpleSeq2Seq (#395)
* Rename EncoderDecoder to SimpleSeq2Seq

This is so that anyone wanting to build a sequence-to-sequence model
in AllenNLP has an obvious simple starting place.  I also copied the
`EncoderDecoder` model into a new `WikiTablesSemanticParser` model,
which we'll be modifying as we add in various components that the
original parser had.  So this effectively takes what was
`EncoderDecoder` and makes two classes, one that is simpler (I made
some very small changes to simplify the code slightly, and added some
comments in the main docstring), and one that will soon get a whole lot
more complicated.

* Remove work in progress from PR to master branch