Got the parser actually training on most of the data
This was a lot of work figuring out why some logical forms wouldn't
parse, and fixing the problems, adding test cases as I went.  I also
found some bugs in the logic for computing new states, so I wrote some
more test cases for the decoder, and fixed the bugs.  This now
successfully reads in the subsampled data and starts training, though
there is still a problematic instance that causes the training to crash
partway through.  Looking into that next...