Train parser on actual wikitables data (#639)
* Trying to get this to run with real data, lots of data loading fixes

* Trying to figure out a few more things...

* Got the parser actually training on most of the data

This was a lot of work figuring out why some logical forms wouldn't
parse, and fixing the problems, adding test cases as I went.  I also
found some bugs in the logic for computing new states, so I wrote some
more test cases for the decoder, and fixed the bugs.  This now
successfully reads in the subsampled data and starts training, though
there is still a problematic instance that causes the training to crash
partway through.  Looking into that next...

* Make it so that parsing logical forms doesn't change world state

* Fixed a bug with batched input

* Add a parse accuracy metric, with code provided by Oyvind

* Pylint

* Fix dataset reader test

* Fixed tests

* Fix typo

* Removed unused import

* Fix mypy

* Remove now-unnecessary large sample file

* Address PR feedback

* Add a TODO about sorting the LFs