Implement a sqlite3 index for cached datasets
This commit provides a working implementation of a sqlite3 index for
mapping URLs to filenames in file_utils.get_from_cache(), but it is not
complete. It still needs some more unit testing for some corner cases,
such as when a file indexed in the database does not exist on the
filesystem (it should be removed from the database).

This change is necessary for #1221, where the long base64-encoded
filenames can reach the maximum filename length (255 generally, but
the filesystem is encrypted with encryptfs it is 143).