Implement a sha256 hash-based file caching method
Previously, remote dataset urls were base64-encoded to generate a
filename for a local cached version, but this caused issues when the
urls were long, as the resulting filenames grew to be greater than
the maximum filename length (255 generally, but only 143 if encryptfs
is used to encrypt a filesystem). This commit replaces the base64
encoding with a sha256 hash (which is always 64 characters; if an
ETag is used, which concatenates the url hash with the ETag hash
delimited by a ., the total is 129).

Resolves #1221