Read and parse the file first line without re-opening the file.
The function read_num_pretrained_tokens_if_present(uri) parsed
the first line of the file to see if it was a header. This was
done just before re-opening the file and reading it entirely.

With tar files this is slow (no problem with zip), because the file
gets entirely extracted two times.

I could have just make read_num_pretrained_tokens_if_present accept
an opened file, but because the file is not seekable nor it's easy
to make an independent copy of the iterator I preferred another
approach.