Fix chdir in ModelTestCase breaking downstream models (#1418)
ModelTestCase has a few `chdir`s since after we moved the tests into the module, all the relative paths in the test fixtures were broken (since they assumed that we were running the tests in the same directory as the `tests` folder). I got around this by using `os.chdir`, but did not realize that this would break models that import and use `ModelTestCase` (see #1417 ).

I think a reasonable solution to this is to simply remove all the `chdir` calls and add a symbolic link in the project root called `tests` that links to the tests folder (in the module root). This means that the relative paths in the fixtures don't need to be redone and downstream clients using `ModelTestCase` won't deal with issues from `os.chdir`.