Sql text utils (#1717)
Some utility functions for reading in the text2sql data. These are going to be iteratively improved (e.g there is some deduplication that can be done when reading the data, etc), but this is functional so I thought improve on it iteratively afterward.

This is abstracted from a specific data reader because there may well be several dataset readers associated with these datasets, for baselines etc.