allow zero-padded sequences
Up to now, total_timesteps is defined as the size of the second dimension (time-dimension). This makes sense as long as the longest sequence has exactly the same size. In most cases this is a valid assumption, but it does not allow shorter sequences. For example, if a single batch only contains shorter messages, one would have to use narrow to cut them down. However, it would be easier to just use `batch_lenghts.max()` instead of `inputs.size()[1]`.