Get rid of hardcoded Vocabulary class name.
In case of inheriting Vocabulary, the method `from_params` calls the
same method of the registered class if `type` is given. If
`from_params` is not overridden (even if it is), the default
implementation calls hardcoded `Vocabulary.from_instances` and
`Vocabulary.from_files` which makes overriding these methods in
subclass unreachable.