Fix an f-formatted string (#3483)
There is a `KeyError` message in `pretrained_transformer_indexer.py` that appears to be an f-formatted string with broken syntax. This resulted in the literal string `token.text` being printed when the error is raised, instead of the value of the `text` attribute of `token`.

I simply fixed the syntax, so now this error message will print the value of `token.text` to screen when the error is raised.