Expand documentation of modules.conditional_random_field
This PR attempts to expand the documentation the ConditionalRandomField code. Specifically I found the details around the mask and start-end tokens to be confusing on first reading. Hopefully this makes it bit more explicit. 

This is a simple docs-only change. In future might come back and try clean up the code a bit / improve performance (The CRF layer seems to add a pretty non-trivial overhead to training time. Given the sequential loops and lots of pointwise-ops, it seems like a prime use case of TorchScript, but currently it seems the code would need some changes to be compatible.)

Feedback welcome.