Update vocabulary load to a system-agnostic newline (#4342)
* Update vocabulary load to a system-agnostic newline

Hello, 
I had a problem about training a model on a Linux machine and loading on a Windows machine.  The error was:
AssertionError: OOV token not found!

After some debugging I found out that during the vocabulary loading, it was splitting by '\n', where this can cause a difference between Linux and Windows. This PR change the split to OS agnostic method of new-line splitting.

* Use a regex because the splitlines algo split on tabulation chars

* Use a regex because the splitlines algo split on tabulation chars

* Added to changelog

* Added to changelog

* Use a pre-compiled regex

Co-authored-by: Bruno Cabral <bruno@potelo.com.br>