Remove upper bounds for dependencies in requirements.txt
The stated reason for this pinning was to prevent CI breakage, and there
was a bot to help upgrade dependencies automatically. However, now that
the library is being retired and Depdendabot was, this upgrade process
is no longer happening.

Removing these upper bounds will make using the library in the future
much easier. To prevent CI breakage after upper bound removal, they can
be kept in a constraints file for use in CI. It would be up to end users
in the future to limit dependency conflicts if they encounter any.

I have checked the history, and as far as I can tell, none of these
upper pins exist to prevent a known breakage. Therefore, I moved them
all to the constraints file.